refactor: Sidebar 페이지 컴포넌트에서 루티 부분을 컴포넌트로 분리
