+++ frontend/src/layouts/PlaceList/PlaceList.tsx
@@ -1,33 +1,12 @@
 import Flex from '@/@common/components/Flex/Flex';
 import Text from '@/@common/components/Text/Text';
 import { PlaceCard } from '@/domains/places/components/PlaceCard/PlaceCard';
-import { Routie } from '@/domains/routie/types/routie.types';
 
 import gridContainerStyle from './PlaceList.styles';
 import { usePlaceListContext } from './contexts/PlaceListContext';
-import { usePlaceList } from './hooks/usePlaceList';
 
-interface PlaceListProps {
-  routiePlaces: Routie[];
-  setRoutiePlaces: React.Dispatch<React.SetStateAction<Routie[] | undefined>>;
-  onRoutieDataChange?: () => Promise<void>;
-}
-
-const PlaceList = ({
-  routiePlaces,
-  setRoutiePlaces,
-  onRoutieDataChange,
-}: Omit<PlaceListProps, 'places'>) => {
+const PlaceList = () => {
   const { placeList, refetchPlaceList, handleDelete } = usePlaceListContext();
-
-  const placeCardList = usePlaceList({
-    places: placeList,
-    onDelete: handleDelete,
-    onPlaceChange: refetchPlaceList,
-    routiePlaces,
-    setRoutiePlaces,
-    onRoutieDataChange,
-  });
 
   return (
     <Flex
@@ -41,7 +20,7 @@ const PlaceList = ({
     >
       <Text variant="title">장소 목록</Text>
       <div css={gridContainerStyle}>
-        {placeCardList.map(({ id, onSelect, selected, ...rest }) => (
+        {placeList.map(({ id, onSelect, selected, ...rest }) => (
           <PlaceCard
             id={id}
             key={id}
