+++ frontend/src/layouts/Sidebar/Sidebar.tsx
@@ -1,19 +1,16 @@
-import { useEffect, useMemo, useState } from 'react';
+import { useMemo, useState } from 'react';
 
 import Button from '@/@common/components/Button/Button';
 import Flex from '@/@common/components/Flex/Flex';
 import Header from '@/@common/components/Header/Header';
-import Pill from '@/@common/components/Pill/Pill';
 import Text from '@/@common/components/Text/Text';
 import ToggleSwitch from '@/@common/components/ToggleSwitch/ToggleSwitch';
 import AddPlaceModal from '@/domains/places/components/AddPlaceModal/AddPlaceModal';
-import { editRoutieSequence } from '@/domains/routie/apis/routie';
-import RoutiePlaceCard from '@/domains/routie/components/RoutiePlaceCard/RoutiePlaceCard';
+import RoutieSection from '@/domains/routie/components/RoutieSection/RoutieSection';
 import RoutieValidationResultCard from '@/domains/routie/components/RoutieValidationResultCard/RoutieValidationResultCard';
 import RoutieValidationUnavailableCard from '@/domains/routie/components/RoutieValidationUnavailableCard/RoutieValidationUnavailableCard';
+import { useRoutieContext } from '@/domains/routie/contexts/useRoutieContext';
 import { useRoutieValidateContext } from '@/domains/routie/contexts/useRoutieValidateContext';
-import { useCardDrag } from '@/domains/routie/hooks/useCardDrag';
-import { Routes, Routie } from '@/domains/routie/types/routie.types';
 import RoutieSpaceName from '@/domains/routieSpace/components/RoutieSpaceName/RoutieSpaceName';
 import theme from '@/styles/theme';
 
@@ -21,14 +18,8 @@ import { usePlaceListContext } from '../PlaceList/contexts/PlaceListContext';
 
 import TimeInput from './TimeInput';
 
-interface SidebarProps {
-  setRoutiePlaces: React.Dispatch<React.SetStateAction<Routie[] | undefined>>;
-  routiePlaces: Routie[];
-  routes: Routes[] | undefined;
-}
-
-const Sidebar = ({ routiePlaces, setRoutiePlaces, routes }: SidebarProps) => {
-  const getDragProps = useCardDrag(routiePlaces, setRoutiePlaces);
+const Sidebar = () => {
+  const { routes } = useRoutieContext();
   const { refetchPlaceList } = usePlaceListContext();
   const [addModalOpen, setAddModalOpen] = useState(false);
   const {
@@ -37,37 +28,11 @@ const Sidebar = ({ routiePlaces, setRoutiePlaces, routes }: SidebarProps) => {
     isValidateRoutie,
     handleValidateToggle,
     handleTimeChange,
-    validateRoutie,
   } = useRoutieValidateContext();
 
   const totalMovingTime = useMemo(() => {
     return routes?.reduce((acc, cur) => acc + cur.duration, 0) ?? 0;
   }, [routes]);
-
-  useEffect(() => {
-    if (!routiePlaces) return;
-
-    const updated = routiePlaces.map((item, index) => ({
-      ...item,
-      sequence: index + 1,
-    }));
-
-    const isSequenceChanged = routiePlaces.some(
-      (item, index) => item.sequence !== index + 1,
-    );
-
-    const updateRoutiePlaces = async (sortedPlaces: Routie[]) => {
-      setRoutiePlaces(sortedPlaces);
-      await editRoutieSequence(sortedPlaces);
-      await validateRoutie();
-    };
-
-    if (isSequenceChanged) {
-      const sortedPlaces = updated.sort((a, b) => a.sequence - b.sequence);
-
-      updateRoutiePlaces(sortedPlaces);
-    }
-  }, [routiePlaces]);
 
   const openAddModal = () => {
     setAddModalOpen((prev) => !prev);
@@ -75,15 +40,6 @@ const Sidebar = ({ routiePlaces, setRoutiePlaces, routes }: SidebarProps) => {
 
   const closeAddModal = () => {
     setAddModalOpen((prev) => !prev);
-  };
-
-  const handleDelete = (id: number) => {
-    editRoutieSequence(routiePlaces.filter((item) => item.placeId !== id));
-    setRoutiePlaces(routiePlaces.filter((item) => item.placeId !== id));
-  };
-
-  const MOVING_STRATEGY = {
-    DRIVING: '운전',
   };
 
   return (
@@ -143,7 +99,7 @@ const Sidebar = ({ routiePlaces, setRoutiePlaces, routes }: SidebarProps) => {
             padding: '0 1.6rem',
           }}
         >
-          <Text variant="subTitle">마이 루티</Text>
+          <Text variant="subTitle">내 동선</Text>
           <Flex
             direction="column"
             gap={1}
@@ -155,38 +111,7 @@ const Sidebar = ({ routiePlaces, setRoutiePlaces, routes }: SidebarProps) => {
               boxSizing: 'border-box',
             }}
           >
-            {routiePlaces?.map((place, index) => {
-              return (
-                <>
-                  <div key={place.placeId} {...getDragProps(index)}>
-                    <RoutiePlaceCard
-                      placeId={place.placeId}
-                      handleDelete={() => handleDelete(place.placeId)}
-                      onPlaceChange={refetchPlaceList}
-                    />
-                  </div>
-
-                  {routiePlaces.length - 1 !== index &&
-                    routes &&
-                    routes[index] && (
-                      <Flex key={place.id + index} gap={1}>
-                        <Text variant="description">
-                          {MOVING_STRATEGY[routes[index].movingStrategy]}{' '}
-                          {routes[index].duration}분
-                        </Text>
-                        <Pill type="distance">
-                          <Text
-                            variant="description"
-                            color={theme.colors.purple[400]}
-                          >
-                            {(routes[index].distance / 1000).toFixed(1)}km
-                          </Text>
-                        </Pill>
-                      </Flex>
-                    )}
-                </>
-              );
-            })}
+            <RoutieSection onPlaceChange={refetchPlaceList} />
           </Flex>
         </Flex>
       </Flex>
